build: Refactor projen settings
