+++ projenrc/config/browsersList.ts
@@ -1,0 +1,20 @@
+const browsersList = {
+  production: [
+    ">0.2%",
+    "not dead",
+    "not ie 11",
+    "not chrome < 51",
+    "not safari < 10",
+    "not android < 51",
+  ],
+  development: [
+    ">0.2%",
+    "not dead",
+    "not ie 11",
+    "not chrome < 51",
+    "not safari < 10",
+    "not android < 51",
+  ],
+};
+
+export default browsersList;
